+++ pyspelling/filters/context.py
@@ -1,7 +1,6 @@
 """Text parser."""
 from .. import filters
 from .. import util
-import codecs
 import re
 from collections import OrderedDict
 
@@ -77,7 +76,7 @@ class ContextFilter(filters.Filter):
     def filter(self, source_file, encoding):  # noqa A001
         """Parse file."""
 
-        with codecs.open(source_file, 'r', encoding=encoding) as f:
+        with open(source_file, 'r', encoding=encoding, errors='strict') as f:
             text = f.read()
 
         return [filters.SourceText(self._filter(text), source_file, encoding, 'context')]