+++ docs/src/markdown/about/changelog.md
@@ -7,6 +7,7 @@
 -   **NEW**: Drop support for Python 3.8.
 -   **NEW**: Add official support for Python 3.13.
 -   **NEW**: Add official support for Python 3.14.
+-   **NEW**: Python filter is now aware of Python 3.14+ t-strings.
 
 ## 2.10